backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-jose[cryptography]==3.3.0
boto3==1.34.100
requests==2.31.0
python-multipart==0.0.9
